# swagger.yaml
openapi: 3.0.0
info:
  title: API ClientHub
  version: 1.0.0
paths:
  /login:
    post:
      summary: Login do usuário
      description: Autentica o usuário e retorna um token JWT.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - senha
              properties:
                email:
                  type: string
                  example: daniel@gmail.com
                senha:
                  type: string
                  example: 123456
      responses:
        "200":
          description: Login realizado com sucesso
          content:
            application/json:
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "401":
          description: Credenciais inválidas
  /cadastro:
    post:
      summary: Cadastro de Usuário
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - email
                - senha
                - tipo_usuario
              properties:
                nome:
                  type: string
                  example: Douglas Silva
                email:
                  type: string
                  example: douglas@gmail.com
                senha:
                  type: string
                  example: "12345678"
                tipo_usuario:
                  type: string
                  example: usuario
              responses:
                "201":
                  description: Usuário cadastrado com sucesso!
  /usuario/{usuarioId}/atualizar:
    put:
      summary: Atualizar dados do usuário
      description: Atualiza informações básicas do usuário. O próprio usuário pode
        atualizar seus dados. Admin pode atualizar qualquer usuário.
      tags:
        - Usuários
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: usuarioId
          required: true
          schema:
            type: string
          description: ID do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  example: Daniel Scheffel
                email:
                  type: string
                  example: daniel@email.com
                senha:
                  type: string
                  example: 87654321
      responses:
        "200":
          description: Usuário atualizado com sucesso
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "404":
          description: Usuário não encontrado
  /usuario/{usuarioId}/status:
    put:
      summary: Atualizar status do usuário
      description: Ativa ou desativa um usuário. Apenas administradores.
      tags:
        - Usuários
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: usuarioId
          required: true
          schema:
            type: string
          description: ID do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ativo
              properties:
                ativo:
                  type: boolean
                  example: false
      responses:
        "200":
          description: Status do usuário atualizado
        "401":
          description: Não autenticado
        "403":
          description: Apenas administradores
        "404":
          description: Usuário não encontrado
  /usuarios/{usuarioId}/reatribuir-clientes:
    patch:
      summary: Reatribuir clientes de um usuário
      description: Reatribui todos os clientes de um usuário para outro. Apenas
        administradores.
      tags:
        - Usuários
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: usuarioId
          required: true
          schema:
            type: string
          description: ID do usuário origem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - novoUsuarioId
              properties:
                novoUsuarioId:
                  type: string
                  example: b7c1e5f2-1234-4a9d-9a88-999999999999
      responses:
        "200":
          description: Clientes reatribuídos com sucesso
        "401":
          description: Não autenticado
        "403":
          description: Apenas administradores
        "404":
          description: Usuário não encontrado
  /cadastro/cliente:
    post:
      summary: Cadastrar novo cliente
      description: Cria um novo cliente vinculado ao usuário autenticado.
      tags:
        - Clientes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - email
                - contato
                - empresa
                - tipo_cliente
                - cpf
                - cnpj
                - origem
                - status
                - cidade
                - estado
                - observacao
              properties:
                nome:
                  type: string
                  example: João da Silva
                email:
                  type: string
                  example: joao@email.com
                contato:
                  type: string
                  example: "11999999999"
                empresa:
                  type: string
                  example: Empresa XPTO
                tipo_cliente:
                  type: string
                  example: PF
                  description: PF, PJ ou MEI
                cpf:
                  type: string
                  example: "12345678900"
                cnpj:
                  type: string
                  example: "12345678000199"
                origem:
                  type: string
                  example: site
                status:
                  type: string
                  example: novo
                  description: novo, contatado, negociacao, fechado ou perdido
                cidade:
                  type: string
                  example: São Paulo
                estado:
                  type: string
                  example: SP
                observacao:
                  type: string
                  example: Cliente interessado em plano anual
      responses:
        "201":
          description: Cliente cadastrado com sucesso
        "400":
          description: Erro de validação
        "401":
          description: Não autenticado
components: {}
tags:
  - name: Auth
    description: Autenticação e autorização de usuários e criação de usuário
  - name: Clientes
    description: Funcionamento dos usuários ao clientes